login validation

+++ src/filters/date.filter.js
@@ -1,11 +1,9 @@
 export default function dateFilter(value, format = 'date') {
-  const options = {
-
-  }
+  const options = {}
 
   if (format.includes('date')) {
     options.day = '2-digit';
-    options.month = 'long';
+    options.month = 'short';
     options.year = 'numeric';
   }
